pyproject.toml
[tool.poetry]
name = "saas"
version = "0.1.0"
description = ""
authors = ["dntrieu-nguyen <nguyen.dang@rikai.technology>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
python-dotenv = "^1.0.1"
markdown = "^3.7"
django-filter = "^24.3"
django-guardian = "^2.4.0"
django-rest-swagger = "^2.2.0"
dj-database-url = "^2.3.0"
mysqlclient = "^2.2.5"
pillow = "^11.0.0"
drf-spectacular = {extras = ["sidecar"], version = "^0.27.2"}
django = "^5.1.3"
firebase-admin = "^6.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
